feat(xiuxian): 优化修炼系统并调整游戏数值
- 提高基础修为和灵石获取量
- 实现无限时长修炼，由用户自行决定结束时间
- 优化任务系统，支持跨天任务和时间更长的任务
- 修复部分指令无法使用的问题
- 优化数据保存和加载逻辑

+++ utils.py
@@ -106,8 +106,8 @@ class XiuXianUtils:
         Returns:
             包含奖励信息的字典
         """
-        # 计算基础值
-        base = base_value * multiplier
+        # 计算基础值（确保是整数）
+        base = int(base_value * multiplier)
         # 添加等级加成
         level_bonus = level * level_factor
         # 添加随机波动
